main now takes in 2 arguments, more in desc
main now takes 2 arguments, opens a file using the first item reads it's content and puts it into an array of strings called strings. then takes the second argument and uses it to create a new file and puts the last line read from the previous file into it. There is a commented out line that calls the parser and passes it the array of strings called strings and whatever it returns is what gets added to the file instead. but that can be changed later. Next I will try and catch errors while inputting command line arguments like if they enter more or less than 2 command line arguments and later i will try and error check if the input file doesn't exist