Added ConstraintTup
Added CosntraintTup which is to be populated by the parser. It will also be returned by the parser. The error variable inside the ConstraintTup is to be used for any parsing errors. Main will check that it is of length zero before passing what the parser returned to the BnB function. if the length of error in ConstraintTup that was returned by the Parser is more than zero the program will output the first string in error.